Rename Tests folder to Analyzers
