--- Capgemini.CSharpCodeAnalysis/Capgemini.CodeAnalysis.CoreAnalysers/Capgemini.CodeAnalysis.CoreAnalysers/CodeFixes/NamingConventionCodeFixProvider.cs.bak
@@ -1,72 +1,0 @@
-﻿//using System.Composition;
-//using System.Collections.Immutable;
-//using System.Linq;
-//using System.Threading;
-//using System.Threading.Tasks;
-//using Microsoft.CodeAnalysis;
-//using Microsoft.CodeAnalysis.CodeFixes;
-//using Microsoft.CodeAnalysis.CodeActions;
-//using Microsoft.CodeAnalysis.CSharp;
-//using Microsoft.CodeAnalysis.Rename;
-//using Capgemini.CodeAnalysis.CoreAnalysers.Models;
-//using Capgemini.CodeAnalysis.CoreAnalysers.Extensions;
-
-//namespace Capgemini.CodeAnalysis.CoreAnalysers.CodeFixes
-//{
-//    [ExportCodeFixProvider(LanguageNames.CSharp, Name = nameof(NamingConventionCodeFixProvider)), Shared]
-//    public class NamingConventionCodeFixProvider : CodeFixProvider
-//    {
-//        public const string DiagnosticId = "FU001";
-
-//        public sealed override ImmutableArray<string> FixableDiagnosticIds
-//        {
-//            get
-//            {
-//                return ImmutableArray.Create(DiagnosticId);
-//            }
-//        }
-
-//        public sealed override FixAllProvider GetFixAllProvider()
-//        {
-//             See https://github.com/dotnet/roslyn/blob/master/docs/analyzers/FixAllProvider.md for more information on Fix All Providers
-//            return WellKnownFixAllProviders.BatchFixer;
-//        }
-
-//        public sealed override async Task RegisterCodeFixesAsync(CodeFixContext context)
-//        {
-//            var root = await context.Document.GetSyntaxRootAsync(context.CancellationToken).ConfigureAwait(false);
-
-//             TODO: Replace the following code with your own analysis, generating a CodeAction for each fix to suggest
-//            var diagnostic = context.Diagnostics.First();
-//            var diagnosticSpan = diagnostic.Location.SourceSpan;
-
-//            // Find the type declaration identified by the diagnostic.
-//            var declaration = root.FindToken(diagnosticSpan.Start).Parent.AncestorsAndSelf().OfType<FieldDeclarationSyntax>().First();
-
-//            // Register a code action that will invoke the fix.
-//            context.RegisterCodeFix(
-//                CodeAction.Create(
-//                    title: "Naming Convention code fix",
-//                    createChangedSolution: c => MakeUppercaseAsync(context.Document, declaration, c),
-//                    equivalenceKey: "Naming Convention code fix"),
-//                diagnostic);
-
-//            var token = root.FindToken(diagnostic.Location.SourceSpan.Start);
-//            context.RegisterCodeFix(
-//                CodeAction.Create("Prepend `_` to field", c => PrependUnderscore(context.Document, token, c), AnalyzerType.NamingConventionAnalyzerId.ToDiagnosticId()),
-//                diagnostic);
-//        }
-
-//        private async Task<Solution> PrependUnderscore(Document document, SyntaxToken declaration, CancellationToken cancellationToken)
-//        {
-//            var textCultureInfo = new CultureInfo("en-us").TextInfo;
-//            var titleCasedValue = textCultureInfo.ToTitleCase(declaration.ValueText);
-//            var camelCasedValue = Char.ToLowerInvariant(titleCasedValue[0]) + titleCasedValue.Substring(1);
-//            var newName = $"_{declaration.ValueText}";
-//            var semanticModel = await document.GetSemanticModelAsync(cancellationToken).ConfigureAwait(false);
-//            var symbol = semanticModel.GetDeclaredSymbol(declaration.Parent, cancellationToken);
-//            var solution = document.Project.Solution;
-//            return await Renamer.RenameSymbolAsync(solution, symbol, newName, solution.Workspace.Options, cancellationToken).ConfigureAwait(false);
-//        }
-//    }
-//}
