+++ Capgemini.CSharpCodeAnalysis/Capgemini.CodeAnalysis.CoreAnalysers/Tests/Capgemini.CodeAnalysis.CoreAnalysers.Test/Analyzers/FileHierarchyAnalyzerTests.cs
@@ -4,7 +4,7 @@ using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using TestHelper;
 
-namespace Capgemini.CodeAnalysis.CoreAnalysers.Test.Tests
+namespace Capgemini.CodeAnalysis.CoreAnalysers.Test.Analyzers
 {
     [TestClass]
     public class FileHierarchyAnalyzerTests : CodeFixVerifier
