+++ Capgemini.CSharpCodeAnalysis/Capgemini.CodeAnalysis.CoreAnalysers/Tests/Capgemini.CodeAnalysis.CoreAnalysers.Test/Analyzers/NamingConventionAnalyzerTests.cs
@@ -4,7 +4,7 @@ using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using TestHelper;
 
-namespace Capgemini.CodeAnalysis.CoreAnalysers.Test.Tests
+namespace Capgemini.CodeAnalysis.CoreAnalysers.Test.Analyzers
 {
     [TestClass]
     public class NamingConventionAnalyzerTests : CodeFixVerifier
@@ -125,8 +125,7 @@ namespace Capgemini.CodeAnalysis.CoreAnalysers.Test.Tests
 
         protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
         {
-            // return new NamingConventionAnalyzer();
-            return null;
+            return new NamingConventionAnalyzer();
         }
     }
 }
